requirements.txt
elastalert
packageurl-python